fix: home not render

+++ server/src/main/java/com/backend/controller/dto/post/PostSummary.java
@@ -13,5 +13,5 @@ public class PostSummary {
   private String content;
   private Instant createdAt;
   private String authorName;
-  private Integer postCount;
+  private Long postCount;
 }
